handle newly implemented Yatir and Redwood landuse types
- before, I replaced the "bare ground tundra" landuse type with
parameters for Yatir. This did not work because WRF, somewhere
internally, set the vegetation fraction for these areas with 0.0. That
makes sense for bare ground tundra but not for Yatir.  I have now
implemented a new landuse type.  This required going back to the index
file for the landuse grib files in WRF's static geography data to
update the maximum landuse index from 21 to 23 (adding two categories
for Yatir and Redwoods).  This commit updates adjust_WRF_inputs.py and
driver_landuse_yatir.py to handle the updated landuse categories.